Move historical context to Pulse and reorder providers
